Add DATA_MSG constant and update methods for data sending
A new constant DATA_MSG was added to the Protocol class, representing a standardized data message format. A corresponding getter method was also included for accessing this constant. In addition, the method for sending data in ProbeHttps.java was updated to utilize MessageBuilder functionality, and error messages have been standardized across the board to output to System.err instead of System.out.