Add new build.gradle for 'domains' module
Created a new 'build.gradle' file for the 'domains' module, defining its basic configuration like plugins, group, version, repositories and dependencies. This includes the usage of 'java' plugin, setting the group as 'org.helmo', and the version as '1.0-SNAPSHOT'. It also declares the 'mavenCentral()' as the repository and adds Junit dependencies for testing.