Implement status checking and enhance data collection
This update introduced a new "STATUSOF" command type that enables the checking of the status of a given command. It further refined the data collection process in both ProbeImapRunable and ProbeSnmpRunable classes, enhancing error handling and adapting responses based on the incoming data. Additionally, changes ensured accurate messages with status updates are sent only when a change in the data status occurs.