Implement data encryption and update test cases
Added AES encryption for data exchanged in the monitor daemon and probe. This was done by introducing the AesEncryption class, encrypting the message before sending it and then decrypting it upon receipt. The encryption keys were updated in the configurations. The ImapBuilder class was updated to accept decrypted credentials. Test cases related to probes were updated to fit these changes. Code handling IMAP was updated for better data collection.