Refactor MonitorDaemon and Probe classes
This commit includes several refactoring changes in the MonitorDaemon and Probe classes. It replaces the processMultiCastMessage with handleProbeMessage method name for clarity. It also changes the way of handling responses from probes and modifies PrintWriter and BufferedWriter implementation to communicate with probes. Lastly, it includes adjustments in the JSON configurations file.